auth validation in ImageServiceImpl

+++ src/main/java/com/example/evently/services/image/ImageServiceImpl.java
@@ -1,5 +1,6 @@
 package com.example.evently.services.image;
 
+import com.example.evently.auth.facade.AuthFacade;
 import com.example.evently.dto.output.Message;
 import com.example.evently.dto.image.ImageReqDelete;
 import com.example.evently.dto.image.ImageRes;
@@ -7,6 +8,7 @@ import com.example.evently.exceptions.BadReqEx;
 import com.example.evently.exceptions.NotFoundEx;
 import com.example.evently.mappers.ImageMapper;
 import com.example.evently.models.Image;
+import com.example.evently.models.User;
 import com.example.evently.repositories.ImageRepository;
 import com.example.evently.services.event.EventService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,20 +20,31 @@ import java.awt.image.BufferedImage;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
 @Service
 public class ImageServiceImpl implements ImageService{
 
     ImageRepository imageRepository;
     CloudinaryService cloudinaryService;
-
     EventService eventService;
+    AuthFacade authFacade;
 
     @Autowired
-    public ImageServiceImpl(ImageRepository imageRepository, CloudinaryService cloudinaryService, EventService eventService) {
+    public ImageServiceImpl(ImageRepository imageRepository,
+                            CloudinaryService cloudinaryService,
+                            EventService eventService,
+                            AuthFacade authFacade) {
         this.imageRepository = imageRepository;
         this.cloudinaryService = cloudinaryService;
         this.eventService = eventService;
+        this.authFacade = authFacade;
+    }
+
+    private User getAuth(){
+        Optional<User> auth = authFacade.getAuthUser();
+        if(auth.isEmpty()) throw new NotFoundEx("User Not Found", "U-404");
+        return auth.get();
     }
 
     @Override
@@ -62,8 +75,10 @@ public class ImageServiceImpl implements ImageService{
     @Override
     public ImageRes upload(MultipartFile multipartFile, Long id) throws IOException {
         var event = eventService.getCompleteEventById(id);
+        if(event.getPublisher() != this.getAuth() && !authFacade.isAdmin())
+            throw new BadReqEx("Only event publisher can upload images!", "I-001");
         BufferedImage bufferedImage = ImageIO.read(multipartFile.getInputStream());
-        if(bufferedImage == null) throw new BadReqEx("Invalid Image", "I-001");
+        if(bufferedImage == null) throw new BadReqEx("Invalid Image", "I-002");
         Map result = cloudinaryService.upload(multipartFile);
         Image image = new ImageMapper().mapCloudinaryResultToImage(result, event);
         imageRepository.save(image);
@@ -73,6 +88,8 @@ public class ImageServiceImpl implements ImageService{
     @Override
     public Message deleteById(Long id) throws IOException {
         Image image = this.findById(id);
+        if(image.getEvent().getPublisher() != this.getAuth() && !authFacade.isAdmin())
+            throw new BadReqEx("Only event publisher can delete its images!", "I-001");
         Map result = cloudinaryService.delete(image.getImgId());
         //validar amb result?
         imageRepository.delete(image);
@@ -82,6 +99,8 @@ public class ImageServiceImpl implements ImageService{
     @Override
     public Message deleteByUrl(ImageReqDelete req) throws IOException {
         Image image = this.findByUrl(req.getUrl());
+        if(image.getEvent().getPublisher() != this.getAuth() && !authFacade.isAdmin())
+            throw new BadReqEx("Only event publisher can delete its images!", "I-001");
         cloudinaryService.delete(image.getImgId());
         imageRepository.delete(image);
         return new Message("Image deleted!");
